+++ package.json
@@ -19,6 +19,7 @@
 		"@sveltejs/adapter-auto": "^4.0.0",
 		"@sveltejs/kit": "^2.16.0",
 		"@sveltejs/vite-plugin-svelte": "^5.0.0",
+		"@types/jsonwebtoken": "^9.0.9",
 		"autoprefixer": "^10.4.20",
 		"bits-ui": "^0.22.0",
 		"clsx": "^2.1.1",
@@ -38,5 +39,10 @@
 		"typescript": "^5.0.0",
 		"typescript-eslint": "^8.20.0",
 		"vite": "^6.0.0"
+	},
+	"dependencies": {
+		"jsonwebtoken": "^9.0.2",
+		"jwt-decode": "^4.0.0",
+		"safe-buffer": "^5.2.1"
 	}
 }
