+++ src/routes/children/+page.svelte
@@ -1,0 +1,1 @@
+<div>CHILDREN!</div>
