+++ src/routes/classes/+page.svelte
@@ -1,0 +1,1 @@
+<div>CLASSES!</div>
