+++ src/routes/collections/+page.svelte
@@ -1,0 +1,4 @@
+<script lang="ts">
+</script>
+
+<div>COLLECTIONS!</div>
