+++ src/routes/reports/+page.svelte
@@ -1,0 +1,1 @@
+<div>REPORTS</div>
